requirements.txt
langchain>=0.3.0
langgraph>=0.2.0
pydantic>=2.7.0
python-dotenv>=1.0.1
requests>=2.32.0
pyyaml>=6.0.1
jinja2>=3.1.4
